refactor(async-select-input): migrate to ts
